requirements.txt
selenium
sqlalchemy
schedule
flask
python-dotenv
